requirement.txt
django
django-crispy-forms
opencv-python
qrcode
django-rest-framework
pillow